original_requirements.txt
absl-py==0.12.0
astor==0.8.1
attrs==20.3.0
backcall==0.2.0
backports.functools-lru-cache==1.6.3
bleach==3.3.0
brotlipy==0.7.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cloudpickle==1.6.0
colorama==0.4.4
cryptography==3.4.7
cycler==0.10.0
cytoolz==0.11.0
dask==2021.4.0
decorator==5.0.5
defusedxml==0.7.1
dgl-cu101==0.6.1
docopt==0.6.2
entrypoints==0.3
fsspec==0.9.0
future==0.18.2
gast==0.2.2
gitdb==4.0.7
GitPython==3.1.14
google-auth==1.28.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.36.1
h5py==2.9.0
idna==2.8
imageio==2.9.0
importlib-metadata==3.10.0
ipykernel==5.1.2
ipython==7.7.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.3
joblib==1.0.1
jsonpickle==1.5.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
locket==0.2.0
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.1.0
mistune==0.8.4
munch==2.5.0
nbconvert==5.6.1
nbformat==5.1.3
networkx==2.3
notebook==6.0.0
numpy==1.18.2
oauthlib==3.1.0
ogb==1.1.1
olefile==0.46
opencv-python==4.2.0.32
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.2
parso==0.7.1
partd==1.1.0
patsy==0.5.1
pexpect==4.8.0
pgmpy==0.1.11
pickleshare==0.7.5
Pillow==7.1.1
plotly==4.1.1
prometheus-client==0.10.0
prompt-toolkit==2.0.10
protobuf==3.15.7
ptyprocess==0.7.0
py-cpuinfo==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.8.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.0
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
sacred==0.8.1
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.0
seaborn==0.11.1
Send2Trash==1.5.0
six==1.15.0
smmap==4.0.0
statsmodels==0.12.2
tensorboard==1.14.0
tensorboardX==1.8
tensorflow-estimator==2.1.0
tensorflow-gpu==2.1.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
toolz==0.11.1
torch==1.7.1+cu101
torchaudio==0.7.2
torchvision==0.8.2+cu101
tornado==6.1
tqdm==4.35.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.25.11
Wand==0.5.8
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.4.1